autotranslate: translate example .scala identifiers via shared CodeGlossary
Swedish identifiers in \scalainputlisting'd example programs (vego1-4.scala etc.)
rendered Swedish in the English build: the mirror ran Code.translate (comments +
strings only, identifiers verbatim) and the inline \ifswedish clamps can't reach
external .scala files. So compendium/slides chapter 10 showed `object
exempelVego1`, `trait Grönsak`, …

- Promote the ratified apply-b0d glossary to a production module
  autotranslate/CodeGlossary.scala (single source of truth; apply-b0d aliases it).
- Mirror renames example .scala/.java identifiers via CodeGlossary.renderCodeIds:
  token-exact identifier rename applied ONLY in code regions (string/char literals
  and //,/* */ comments copied verbatim). Strings/comments are left to
  Code.translate + Overrides, so nothing half-translates and a corpus-wide str
  replace can't misfire. render() (id + str, whole span) is retained for the
  context-scoped slide clamps only.
- Knowledge strings the model gets wrong moved from CodeGlossary.str to
  Overrides.scala (whole-unit match, no substring problem): Skållas.->Blanched.,
  Skalas med skalare.->Peeled with a peeler., Antal skördade/ätvärda grönsaker:.
  Code.translate gains an `alsoTranslate` gate so a no-åäö Swedish string the
  swedishish heuristic misses still routes to the Override (English strings aren't
  Override keys, so they still never reach the model).
- Add vego wrapper object names exempelVego1..4 -> exampleVego1..4.

Verified: --verify-examples COMPILE OK; mirror regenerates vego1.scala as object
exampleVego1 / trait Vegetable / class Cucumber(var weight) / isPeeled, with
println("Peeled with a peeler.") and println("Blanched.") via Overrides;
vegomatch.scala prints "Number of harvested vegetables:". 0 model calls.
